+++ CHANGELOG.md
@@ -6,6 +6,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## Unreleased
 
+### Added
+- Client APIs to provide Admin functionality [#541](https://github.com/rokwire/groups-building-block/issues/541)
+
+
 ## [1.57.1] - 2025-02-03
 ### Changed
 - Fix consolidate the information, and make it accessible with a single API call [#538](https://github.com/rokwire/groups-building-block/issues/538)